requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
redis==5.2.1
pydantic==2.11.3
pydantic-settings==2.9.1
argon2-cffi==23.1.0
PyJWT==2.10.1
python-dotenv==1.1.0
pydantic[email]
greenlet